no functional code change: rename dc_mimefactory to mimefactory and move some functions to become MimeFactory methods
