+++ src/e2ee.rs
@@ -23,12 +23,12 @@ use mmime::{mailmime_substitute, MAILIMF_NO_ERROR, MAIL_NO_ERROR};
 use crate::aheader::*;
 use crate::config::Config;
 use crate::context::Context;
-use crate::dc_mimefactory::*;
 use crate::dc_mimeparser::*;
 use crate::dc_tools::*;
 use crate::error::*;
 use crate::key::*;
 use crate::keyring::*;
+use crate::mimefactory::MimeFactory;
 use crate::peerstate::*;
 use crate::pgp::*;
 use crate::securejoin::handle_degrade_event;
